+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### 🚀 Added
 
-- TBD
+- ASTVisitor will check for multiple visit methods with the same type annotation and raise an error if found
 
 ### ✨ Changed
 
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### 🐛 Fixed
 
-- TBD
+- ASTVisitor didn't cache bound methods properly, it's fixed now
 
 ### 🔥 Removed